refactor : domain 패키지 추가
