+++ docs/README.md
@@ -20,5 +20,5 @@
 
 ### step5 구현할 기능 목록
 
-- [ ] MVC 패턴 기반으로 패키지 분리
+- [x] MVC 패턴 기반으로 패키지 분리
 - [ ] 테스트하기 힘든 부분 분리하여 단위 테스트 늘리기
