+++ src/main/java/step3/CarRacingController.java
@@ -1,5 +1,7 @@
 package step3;
 
+import step3.domain.CarRacingService;
+import step3.domain.MoveStrategy;
 import step3.dto.CarStatusDto;
 import step3.dto.RacingResultDto;
 import step3.view.ResultView;
