+++ src/main/java/step3/domain/AlwaysMoveStrategy.java
@@ -1,4 +1,4 @@
-package step3;
+package step3.domain;
 
 public class AlwaysMoveStrategy implements MoveStrategy {
