+++ src/main/java/step3/domain/Car.java
@@ -1,4 +1,4 @@
-package step3;
+package step3.domain;
 
 public class Car {
 
@@ -6,12 +6,12 @@ public class Car {
     private int location;
     private MoveStrategy moveStrategy;
 
-    Car(MoveStrategy moveStrategy) {
+    public Car(MoveStrategy moveStrategy) {
         this.location = 0;
         this.moveStrategy = moveStrategy;
     }
 
-    Car(String name, MoveStrategy moveStrategy) {
+    public Car(String name, MoveStrategy moveStrategy) {
         this.name = new CarName(name);
         this.location = 0;
         this.moveStrategy = moveStrategy;
