+++ src/main/java/step3/domain/CarName.java
@@ -1,4 +1,4 @@
-package step3;
+package step3.domain;
 
 public class CarName {
