+++ src/test/java/step3/AlwaysMoveStrategy.java
@@ -1,5 +1,7 @@
 package step3;
 
+import step3.domain.MoveStrategy;
+
 public class AlwaysMoveStrategy implements MoveStrategy {
 
     @Override
