+++ src/test/java/step3/RacingCarTest.java
@@ -3,6 +3,7 @@ package step3;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import step3.domain.*;
 
 import java.util.Arrays;
 import java.util.List;
